# packages/cactus-plugin-ledger-connector-corda/src/main/json/openapi.tpl.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hyperledger Cacti Plugin - Connector Corda",
    "description": "Can perform basic tasks on a Corda ledger",
    "version": "2.0.0-rc.3",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "components": {
    "schemas": {
      "SHA256": {
        "description": "SHA-256 is part of the SHA-2 hash function family. Generated hash is fixed size, 256-bits (32-bytes).",
        "externalDocs": {
          "description": "Official Corda documentation of the SHA256 class.",
          "url": "https://api.corda.net/api/corda-os/4.6/html/api/kotlin/corda/net.corda.core.crypto/-secure-hash/-s-h-a256/index.html"
        },
        "type": "object",
        "required": ["bytes", "offset", "size"],
        "properties": {
          "bytes": {
            "type": "string",
            "example": "Vf9MllnrC7vrWxrlDE94OzPMZW7At1HhTETL/XjiAmc="
          },
          "offset": {
            "type": "integer",
            "example": 0
          },
          "size": {
            "type": "integer",
            "example": 32
          }
        }
      },
      "CordappInfo": {
        "type": "object",
        "externalDocs": {
          "description": "Official Corda Documentation of the class CordappInfo",
          "url": "https://api.corda.net/api/corda-os/4.6/html/api/kotlin/corda/net.corda.core.cordapp/-cordapp-info/index.html"
        },
        "description": "A CordappInfo describes a single CorDapp currently installed on the node",
        "required": [
          "jarHash",
          "licence",
          "minimumPlatformVersion",
          "name",
          "shortName",
          "targetPlatformVersion",
          "type",
          "vendor",
          "version"
        ],
        "properties": {
          "jarHash": {
            "type": "object",
            "format": "SHA256",
            "description": "The hash of the JAR file that defines this CorDapp",
            "$ref": "#/components/schemas/SHA256"
          },
          "licence": {
            "type": "string",
            "description": "The name of the licence this CorDapp is released under"
          },
          "minimumPlatformVersion": {
            "type": "integer",
            "description": "The minimum platform version the node must be at for the CorDapp to run"
          },
          "name": {
            "type": "string",
            "description": "The name of the JAR file that defines the CorDapp"
          },
          "shortName": {
            "type": "string",
            "description": "The name of the CorDapp"
          },
          "targetPlatformVersion": {
            "type": "integer",
            "description": "The target platform version this CorDapp has been tested against"
          },
          "type": {
            "type": "string",
            "description": "A description of what sort of CorDapp this is - either a contract, workflow, or a combination."
          },
          "vendor": {
            "type": "string",
            "description": "The vendor of this CorDapp"
          },
          "version": {
            "type": "string",
            "description": "The version of this CorDapp"
          }
        }
      },
      "NodeDiagnosticInfo": {
        "type": "object",
        "externalDocs": {
          "description": "Official Corda Documentation of the class NodeDiagnosticInfo",
          "url": "https://api.corda.net/api/corda-os/4.6/html/api/kotlin/corda/net.corda.core.node/-node-diagnostic-info/index.html"
        },
        "description": "A NodeDiagnosticInfo holds information about the current node version.",
        "required": [
          "cordapps",
          "platformVersion",
          "revision",
          "vendor",
          "version"
        ],
        "properties": {
          "cordapps": {
            "type": "array",
            "description": "A list of CorDapps currently installed on this node",
            "items": {
              "$ref": "#/components/schemas/CordappInfo"
            },
            "minItems": 0,
            "maxItems": 4096
          },
          "platformVersion": {
            "type": "integer",
            "description": "The platform version of this node. This number represents a released API version, and should be used to make functionality decisions (e.g. enabling an app feature only if an underlying platform feature exists)"
          },
          "revision": {
            "type": "string",
            "description": "The git commit hash this node was built from"
          },
          "vendor": {
            "type": "string",
            "description": "The vendor of this node"
          },
          "version": {
            "type": "string",
            "description": "The current node version string, e.g. 4.3, 4.4-SNAPSHOT. Note that this string is effectively freeform, and so should only be used for providing diagnostic information. It should not be used to make functionality decisions (the platformVersion is a better fit for this)."
          }
        }
      },
...
      "ListFlowsV1Request": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "filter": {
            "type": "string"
          }
        }
      },
      "ListFlowsV1Response": {
        "type": "object",
        "required": ["flowNames"],
        "properties": {
          "flowNames": {
            "type": "array",
            "description": "An array of strings storing the names of the flows as returned by the flowList Corda RPC.",
            "example": [
              "net.corda.finance.flows.CashExitFlow",
              "net.corda.finance.flows.CashIssueAndPaymentFlow",
              "net.corda.finance.flows.CashIssueFlow",
              "net.corda.finance.flows.CashPaymentFlow",
              "net.corda.finance.internal.CashConfigDataFlow",
              "net.corda.samples.obligation.flows.IOUIssueFlow",
              "net.corda.samples.obligation.flows.IOUSettleFlow",
              "net.corda.samples.obligation.flows.IOUTransferFlow",
              "net.corda.samples.obligation.flows.SelfIssueCashFlow"
            ],
            "items": {
              "type": "string",
              "minItems": 0,
              "maxItems": 10000000
            },
            "default": []
          }
        }
      },
...
      "DiagnoseNodeV1Request": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "nodeIds": {
            "type": "array",
            "description": "Optional property specifying which Corda Node should be the one being diagnosed in case the Connector has multiple connections established for different nodes (which is not yet a supported feature, but we want to keep this possibility open for the future).",
            "minItems": 0,
            "maxItems": 1024,
            "default": [],
            "items": {
              "type": "string",
              "nullable": false
            }
          }
        }
      },
      "DiagnoseNodeV1Response": {
        "type": "object",
        "required": ["nodeDiagnosticInfo"],
        "properties": {
          "nodeDiagnosticInfo": {
            "$ref": "#/components/schemas/NodeDiagnosticInfo"
          }
        }
      },
...
      "ListCpiV1Response": {
        "type": "object",
        "required": [
          "hash",
          "id",
          "libraries",
          "mainBundle",
          "timestamp",
          "type"
        ],
        "properties": {
          "cpis": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "cpiFileChecksum": {
                  "type": "string",
                  "nullable": false,
                  "example": "string"
                },
                "cpiFileFullChecksum": {
                  "type": "string",
                  "nullable": false,
                  "example": "string"
                },
                "cpks": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "hash": {
                        "type": "string",
                        "nullable": false,
                        "example": "string"
                      },
                      "id": {
                        "$ref": "#/components/schemas/CPIIDV1"
                      },
                      "libraries": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "nullable": false,
                          "example": "string"
                        }
                      },
                      "mainBundle": {
                        "type": "string",
                        "nullable": false,
                        "example": "string"
                      },
                      "timestamp": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": false,
                        "example": "2022-06-24T10:15:30Z"
                      },
                      "type": {
                        "type": "string",
                        "nullable": false,
                        "example": "string"
                      }
                    }
                  }
                },
                "groupPolicy": {
                  "type": "string",
                  "nullable": true,
                  "example": "string"
                },
                "id": {
                  "$ref": "#/components/schemas/CPIIDV1"
                },
                "timestamp": {
                  "type": "string",
                  "format": "date-time",
                  "nullable": false,
                  "example": "2022-06-24T10:15:30Z"
                }
              }
            }
          }
        }
      },
      "ListCpiV1Request": {
        "required": ["username", "password", "rejectUnauthorized"],
        "properties": {
          "username": {
            "type": "string",
            "nullable": false
          },
          "password": {
            "type": "string",
            "nullable": false
          },
          "rejectUnauthorized": {
            "type": "boolean",
            "nullable": false
          }
        }
      },
...
      "CPIIDV1": {
        "type": "object",
        "required": ["name", "version"],
        "properties": {
          "name": {
            "type": "string",
            "nullable": false,
            "example": "string"
          },
          "signerSummaryHash": {
            "type": "string",
            "nullable": true,
            "example": "string"
          },
          "version": {
            "type": "string",
            "nullable": false,
            "example": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/list-flows": {
      "post": {
        "operationId": "listFlowsV1",
        "x-hyperledger-cacti": {
          "http": {
            "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/list-flows",
            "verbLowerCase": "post"
          }
        },
        "description": "Responds with a list of the flows on the Corda node.",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListFlowsV1Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListFlowsV1Response"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/diagnose-node": {
      "post": {
        "operationId": "diagnoseNodeV1",
        "x-hyperledger-cacti": {
          "http": {
            "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/diagnose-node",
            "verbLowerCase": "post"
          }
        },
        "description": "Responds with diagnostic information about the Corda node",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnoseNodeV1Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnoseNodeV1Response"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/listCPI": {
      "get": {
        "x-hyperledger-cactus": {
          "https": {
            "verbLowerCase": "get",
            "path": "/api/v1/plugins/@hyperledger/cactus-plugin-ledger-connector-corda/listCPI"
          }
        },
        "operationId": "listCpiV1",
        "summary": "List all CPIs uploaded to the cluster",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListCpiV1Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCpiV1Response"
                }
              }
            }
          }
        }
      }
    },
...
  }
}